Make the broker work with multiple Docker nodes in a Swarm setup
The current broker works fine with a single Docker node, but doesn't
when multiple Docker nodes are deployed in a Swarm setup. The reason
is that it returns the wrong (broker) IP, instead of the container
(Docker node) IP.

The change is constructed in such a way that it works with single
node Docker setups (sharing the machine with the broker) as well as
multi node Docker/Swarm setups (where the Docker nodes run on
separate machines, i.e. other machines than the broker itself).